[FE] refactor: review
