--- frontend/src/components/star/index.ts
@@ -1,1 +1,0 @@
-export {Star} from './star'
